進捗計算修正とInteractorロジックリファクタリング

+++ ResearchXBRL.Application/Usecase/FinancialReports/IAquireFinancialReportsPresenter.cs
@@ -3,7 +3,7 @@ namespace ResearchXBRL.Application.Usecase.FinancialReports
 {
     public interface IAquireFinancialReportsPresenter
     {
-        void Progress(int percentage);
+        void Progress(double percentage);
         void Complete();
         void Error(string message, Exception ex);
     }
